build(env): se sube proyecto casi terminado se termina el crud de productos se tuvo que cambiar storage locla

+++ src/main/java/com/ecommerce/api/dto/response/PaymentsDTO.java
@@ -2,6 +2,7 @@ package com.ecommerce.api.dto.response;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 @Builder
 @Getter
@@ -13,7 +14,7 @@ public class PaymentsDTO {
     private Long orderId;
     private Long userId;
     private String paymentMethod;
-    private Long paymentAmount;
+    private BigDecimal paymentAmount;
     private String paymentStatus;
     private String paymentCurrency;
     private LocalDate paymentDate;
